requirements.txt
beautifulsoup4==4.4.1
lxml==3.4.4
Pillow==3.0.0
requests==2.8.1
selenium==2.48.0
wheel==0.24.0
